% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degToRad.R
\name{degToRad}
\alias{degToRad}
\title{Convert degree angle to radians.}
\usage{
degToRad(angleDeg)
}
\arguments{
\item{angleDeg}{angle in degree.}
}
\value{
Angle in radians.
}
\description{
Convert degree angle to radians.
}
\author{
Ana Casanueva (10.01.2017).
}
