% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewp2hurs.R
\name{dewp2hurs}
\alias{dewp2hurs}
\title{Calculation of relative humidity from temperature and dewpoint temperature.}
\usage{
dewp2hurs(tas, dewp)
}
\arguments{
\item{tas}{vector of temperature in degC.}

\item{dewp}{vector of dewpoint temperature in degC.}
}
\value{
relative humidity in %
}
\description{
Calculation of relative humidity from temperature and dewpoint temperature.
}
\details{
Formulation from Dosseger et al. 1992. Formula 99 in MCH document.
}
\author{
Ana Casanueva (11.08.201)
}
