% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diffusivity.R
\name{diffusivity}
\alias{diffusivity}
\title{Compute the diffusivity of water vapor in air.}
\usage{
diffusivity(Tk, Pair)
}
\arguments{
\item{Tk:}{value of air temperature in Kelvin.}

\item{Pair:}{value of air pressure in hPa.}
}
\value{
diffusivity of water vapor in air, m2/s
}
\description{
Compute the diffusivity of water vapor in air, m2/s
}
\details{
Reference: BSL, page 505.
}
\author{
Ana Casanueva (05.01.2017).
}

