% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discomInd.R
\name{discomInd}
\alias{discomInd}
\title{Calculation of the discomfort index.}
\usage{
discomInd(tas, hurs)
}
\arguments{
\item{tas}{vector of air temperature in degC.}

\item{hurs}{vector of relative humidity in \%.}
}
\value{
Discomfort index in degC.
}
\description{
Calculation of the discomfort index from temperature and relative humidity.
}
\details{
Formula based on air temperature and relative humidity, as it is calculated in Coccolo et al. 2016 and references therein.
}
\examples{
\dontrun{ 
# load the meteorological variables for example data in Salamanca:
data("data_obs") 
di <- discomInd(data_obs$tasmean, hurs=data_obs$hurs)
}

}
\author{
A.Casanueva (22.03.2018).
}
