% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveTemp.R
\name{effectiveTemp}
\alias{effectiveTemp}
\title{Calculation of the effective temperature.}
\usage{
effectiveTemp(tas, hurs, wind)
}
\arguments{
\item{tas}{vector of air temperature in degC.}

\item{hurs}{vector of relative humidity in \%.}

\item{wind}{vector of wind at 10m in m/s.}
}
\value{
Effective temperature in degC.
}
\description{
Calculation of the effective temperature from temperature, relative humidity and wind.
}
\details{
Formula based on air temperature, relative humidity and wind, as it is calculated in Coccolo et al. 2016 and references therein.
}
\examples{
\dontrun{ 
# load the meteorological variables for example data in Salamanca:
data("data_obs") 
et <- effectiveTemp(data_obs$tasmean, hurs=data_obs$hurs, wind=data_obs$wind)
}

}
\author{
A.Casanueva (22.03.2018).
}
