% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_atm.R
\name{emis_atm}
\alias{emis_atm}
\title{Calculate the atmospheric emissivity.}
\usage{
emis_atm(Tk, RH)
}
\arguments{
\item{Tk:}{value of air temperature in Kelvin.}

\item{RH:}{value of relative humidity in fraction.}
}
\value{
atmospheric emissivity.
}
\description{
Calculate the atmospheric emissivity.
}
\details{
Reference: Oke (2nd edition), page 373.
}
\author{
Ana Casanueva (05.01.2017).
}

