% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esat.R
\name{esat}
\alias{esat}
\title{Calculate the saturation vapor pressure (hPa) over water.}
\usage{
esat(Tk)
}
\arguments{
\item{Tk:}{value of air temperature in Kelvin.}
}
\value{
saturation vapor pressure (hPa).
}
\description{
Calculate the saturation vapor pressure (hPa) over water.
}
\details{
Reference: Buck's (1981) approximation (eqn 3) of Wexler's (1976) formulae over liquid water.
}
\author{
Ana Casanueva (05.01.2017).
}
