% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fTg.R
\name{fTg}
\alias{fTg}
\title{Calculation of the globe temperature.}
\usage{
fTg(
  tas,
  relh,
  Pair,
  wind,
  min.speed,
  radiation,
  propDirect,
  zenith,
  SurfAlbedo = 0.4,
  tolerance = 1e-04
)
}
\arguments{
\item{tas}{vector of temperature in degC.}

\item{relh}{vector of relative humidity in \%.}

\item{Pair}{value of air pressure in hPa.}

\item{wind}{vector of wind speed in m/s.}

\item{min.speed}{value of minimum wind speed in m/s.}

\item{radiation}{vector of solar shortwave downwelling radiation in W/m2.}

\item{propDirect}{proportion of direct radiation = direct/(diffuse + direct).}

\item{zenith}{zenith angle in radians.}

\item{SurfAlbedo}{(optional) albedo in the surface. Default: 0.4.}

\item{tolerance}{(optional) tolerance value for the iteration. Default: 1e-4.}
}
\value{
Globe temperature in degC.
}
\description{
Calculation of the globe temperature.
}
\details{
Original fortran code by James C. Liljegren, translated by Bruno Lemke into Visual Basic (VBA).
}
\author{
Ana Casanueva (05.01.2017).
}
