% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fTnwb.R
\name{fTnwb}
\alias{fTnwb}
\title{Calculation of the natural wet bulb temperature.}
\usage{
fTnwb(Ta, Td, relh, Pair, ws, min.speed, solar, propDirect, zenith, irad = 1,
  SurfAlbedo = 0.4, tolerance = 1e-06)
}
\arguments{
\item{irad}{(optional): include radiation (1) or not (irad=0, psychrometric web bulb temp). Default: 1.}

\item{SurfAlbedo}{(optional): albedo in the surface. Default: 0.4.}

\item{tolerance}{(optional): tolerance value for the iteration. Default: 1e-6.}

\item{Ta:}{value of air temperature in degC.}

\item{relh:}{value of relative humidity in \%.}

\item{Pair:}{value of atmospheric pressure in hPa.}

\item{ws:}{value of wind speed in m/s.}

\item{min.speed:}{value of minimum wind speed in m/s.}

\item{solar:}{value of solar shortwave downwelling radiation in W/m2.}

\item{propDirect:}{proportion of direct radiation = direct/(diffuse + direct).}

\item{zenith:}{zenith angle in radians.}
}
\value{
Natural wet bulb globe temperature in degC.
}
\description{
Calculation of the natural wet bulb temperature.
}
\details{
Original fortran code by James C. Liljegren, translated by Bruno Lemke into Visual Basic (VBA).
}
\author{
Ana Casanueva (05.01.2017).
}

