% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/h_cylinder_in_air.R
\name{h_cylinder_in_air}
\alias{h_cylinder_in_air}
\title{Calculate the convective heat transfer coefficient for a long cylinder in cross flow.}
\usage{
h_cylinder_in_air(Tk, Pair, speed, min.speed, diam.wick)
}
\arguments{
\item{Tk:}{value of air temperature in Kelvin.}

\item{Pair:}{value of air pressure in hPa.}

\item{speed:}{value of wind speed in m/s.}

\item{min.speed:}{value of minimum wind speed in m/s.}

\item{diam.wick:}{diameter of the cylinder in m.}
}
\value{
Convective heat transfer coefficient for a long cylinder, W/(m2 K).
}
\description{
Calculate the convective heat transfer coefficient for a long cylinder in cross flow.
}
\details{
Reference: Bedingfield and Drew, eqn 32.
}
\author{
Ana Casanueva (05.01.2017).
}

