% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_evap.R
\name{h_evap}
\alias{h_evap}
\title{Calculate the heat of evaporation, J/(kg K).}
\usage{
h_evap(Tk)
}
\arguments{
\item{Tk:}{value of air temperature in Kelvin.}
}
\value{
Heat of evaporation, J/(kg K).
}
\description{
Calculate the heat of evaporation, J/(kg K), for temperature in the range 283-313 K.
}
\details{
Reference: Van Wylen and Sonntag, Table A.1.1
}
\author{
Ana Casanueva (05.01.2017).
}
