% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_sphere_in_air.R
\name{h_sphere_in_air}
\alias{h_sphere_in_air}
\title{Calculate the convective heat tranfer coefficient for flow around a sphere.}
\usage{
h_sphere_in_air(Tk, Pair, speed, min.speed, diam.globe)
}
\arguments{
\item{Tk}{value of air temperature in Kelvin.}

\item{Pair}{value of air pressure in hPa.}

\item{speed}{value of wind speed in m/s.}

\item{min.speed}{value of minimum wind speed in m/s.}

\item{diam.globe}{diameter of the sphere in m.}
}
\value{
Convective heat tranfer coefficient for flow around a sphere, W/(m2 K).
}
\description{
Calculate the convective heat tranfer coefficient for flow around a sphere.
}
\details{
Reference: Bird, Stewart, and Lightfoot (BSL), page 409.
}
\author{
Ana Casanueva (05.01.2017).
}
