% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatStress.R
\docType{package}
\name{HeatStress}
\alias{HeatStress}
\alias{HeatStress-package}
\title{HeatStress}
\description{
Calculate heat stress indices
}
\details{
The package \code{HeatStress} provides methods to calculate heat stress indices.



The following calculation methods are implemented:
 \itemize{
 
 \item \code{wbt.Stull}: Calculation of the Wet Bulb Temperature (Stull 2011).
 \item \code{wbgt.Bernard}: Calculation of the Wet Bulb Globe Temperature in the shade (Bernard et al. 1999) 
 \item \code{wbgt.Liljegren}: Calculation of the Wet Bulb Globe Temperature in the sun (Liljegren et al. 2008) 
 }
}
