% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{Calculation of the heat index.}
\usage{
hi(tas, hurs)
}
\arguments{
\item{tas}{vector of air temperature in degC.}

\item{hurs}{vector of relative humidity in \%.}
}
\value{
Heat index in degrees Fahrenheit.
}
\description{
Calculation of the heat index from temperature and relative humidity.
}
\details{
Formula based on air temperature and relative humidity, as it is calculated in Buzan et al. 2015, but adapted for degrees Celsius.
}
\examples{
\dontrun{ 
# load the meteorological variables for example data in Salamanca:
data("data_obs") 
heatindex <- hi(data_obs$tasmean, hurs=data_obs$hurs)
}

}
\author{
A.Casanueva (22.03.2018).
}
