% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidex.R
\name{humidex}
\alias{humidex}
\title{Calculation of humidex.}
\usage{
humidex(tas, hurs)
}
\arguments{
\item{tas}{vector of air temperature in degC.}

\item{hurs}{vector of relative humidity in \%.}
}
\value{
Humidex values in degC.
}
\description{
Calculation of humidex from temperature and relative humidity.
}
\details{
Formula based on air temperature and relative humidity, as it is calculated in Buzan 2015 GMD and references therein.
}
\examples{
\dontrun{ 
# load the meteorological variables for example data in Salamanca:
data("data_obs") 
hum <- humidex(data_obs$tasmean, hurs=data_obs$hurs)
}

}
\author{
A.Casanueva (22.03.2018).
}
