% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.leapyear.R
\name{is.leapyear}
\alias{is.leapyear}
\title{Check whether a year is a leap year.}
\usage{
is.leapyear(year)
}
\arguments{
\item{Year}{to be checked.}
}
\value{
logical, TRUE/FALSE.
}
\description{
Check whether a year is a leap year.
}
\author{
Sven Kotlarski (20.12.2016).
}
