% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wbgt.Bernard.R
\docType{data}
\name{"data_wbgt.Bernard"}
\alias{"data_wbgt.Bernard"}
\alias{data_wbgt.Bernard}
\title{Example for Wet Bulb Globe Temperature in the shade (Bernard et al. 1999).}
\format{List of two elements: 
\itemize{
\item data: numeric vector with the WBGT values in degC.
\item Tpwb: numeric vector with the phychrometric wet bulb temperature (Tpwb) in degC.
}}
\usage{
data_wbgt.Bernard
}
\description{
Dataset to check the result of \code{wbgt.Bernard} with the testing data \code{obs_data} 
for summer 2003 and one station (Salamanca, Spain).
}
\keyword{datasets}
