% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radToDeg.R
\name{radToDeg}
\alias{radToDeg}
\title{Convert radian angle to degrees.}
\usage{
radToDeg(angleRad)
}
\arguments{
\item{angleRad:}{angle in radians.}
}
\value{
Angle in degrees.
}
\description{
Convert radian angle to degrees.
}
\author{
Ana Casanueva (10.01.2017).
}

