% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swbgt.R
\name{swbgt}
\alias{swbgt}
\title{Calculation of the simplified wet bulb globe temperature.}
\usage{
swbgt(tas, hurs)
}
\arguments{
\item{tas}{vector of air temperature in degC.}

\item{hurs}{vector of relative humidity in \%.}
}
\value{
Simplified wet bulb globe temperature in degC.
}
\description{
Calculation of the simplified wet bulb globe temperature from temperature and relative humidity.
}
\details{
Formula based on air temperature and relative humidity, as it is calculated in Buzan et al. 2015 GMD with a small correction in the constants from Lemke and Kjellstrom 2012.
}
\examples{
\dontrun{ 
# load the meteorological variables for example data in Salamanca:
data("data_obs") 
swbgt <- swbgt(data_obs$tasmean, hurs=data_obs$hurs)
}

}
\author{
A.Casanueva (22.03.2018).
}
