% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tashurs2vap.pres.R
\name{tashurs2vap.pres}
\alias{tashurs2vap.pres}
\title{Calculation of vapour pressure.}
\usage{
tashurs2vap.pres(tas, hurs)
}
\arguments{
\item{tas}{vector of air temperature in degC.}

\item{hurs}{vector of relative humidity in \%.}
}
\value{
Vapour pressure in hPa.
}
\description{
Calculation of vapour pressure from temperature and relative humidity
}
\details{
Formulation from Dosseger et al. 1992. Formula 16 in MCH document.
}
\examples{
\dontrun{ 
# load the meteorological variables for example data in Salamanca:
data("data_obs") 
vp <- tashurs2vap.pres(data_obs$tasmean, hurs=data_obs$hurs)
}

}
\author{
A.Casanueva (11.08.2016).
}
