% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal_cond.R
\name{thermal_cond}
\alias{thermal_cond}
\title{Calculate the thermal conductivity of air, W/(m K).}
\usage{
thermal_cond(Tk)
}
\arguments{
\item{Tk:}{value of air temperature in Kelvin.}
}
\value{
Thermal conductivity of air, W/(m K).
}
\description{
Calculate the thermal conductivity of air, W/(m K).
}
\details{
Reference: BSL, page 257.
}
\author{
Ana Casanueva (05.01.2017).
}

