% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhi.R
\name{uhi}
\alias{uhi}
\title{Calculation of the universal heat index.}
\usage{
uhi(tas, hurs)
}
\arguments{
\item{tas}{vector of air temperature in degC.}

\item{hurs}{vector of relative humidity in \%.}
}
\value{
Universal heat index in degrees Fahrenheit.
}
\description{
Calculation of the universal heat index from temperature and relative humidity.
}
\details{
Formula based on air temperature and relative humidity, as it is calculated in Buzan et al. 2015, but adapted for degrees Celsius.
}
\examples{
\dontrun{ 
# load the meteorological variables for example data in Salamanca:
data("data_obs") 
hi <- uhi(data_obs$tasmean, hurs=data_obs$hurs)
}

}
\author{
A.Casanueva (22.03.2018).
}
