% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscosity.R
\name{viscosity}
\alias{viscosity}
\title{Compute the viscosity of air, kg/(m s).}
\usage{
viscosity(Tk)
}
\arguments{
\item{Tk:}{value of air temperature in Kelvin.}
}
\value{
viscosity of air, kg/(m s).
}
\description{
Compute the viscosity of air, kg/(m s) given temperature (K).
}
\details{
Reference: BSL, page 23.
}
\author{
Ana Casanueva (05.01.2017).
}
