% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wbgt.Bernard.R
\name{wbgt.Bernard}
\alias{wbgt.Bernard}
\title{Calculation of wet bulb globe temperature, following Bernard's method.}
\usage{
wbgt.Bernard(tas, dewp, tolerance = 1e-06)
}
\arguments{
\item{tolerance}{(optional): tolerance value for the iteration. Default: 1e-6.}

\item{tas:}{vector of air temperature in degC.}

\item{dewp:}{vector of dew point temperature in degC.}
}
\value{
A list of:

$value: wet bulb globe temperature in degC.

$tpwb: phychrometric wet bulb temperature (Tpwb) in degC.
}
\description{
Calculation of wet bulb globe temperature from air temperature and dew point temperature. This corresponds to the implementation for indoors or shadow conditions.
}
\details{
Based on Lemke and Kjellstrom 2012, using the formulation from Bernard et al. 1999.
}
\examples{
\dontrun{
# load the meteorological variables for example data in Salamanca:
data("eca_salam_jja_2003.Rdata")
wbgt.indoors <- wbgt.Bernard(eca_salam_jja_2003$tasmean, eca_salam_jja_2003$dewp)
}
}
\author{
A.Casanueva (24.10.2016).
}

