% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wbgt.Liljegren.R
\name{wbgt.Liljegren}
\alias{wbgt.Liljegren}
\title{Calculation of wet bulb globe temperature, following Liljegren's method.}
\usage{
wbgt.Liljegren(tas, dewp, wind, radiation, dates, lon, lat, tolerance = 1e-04,
  noNAs = FALSE)
}
\arguments{
\item{tolerance}{(optional): tolerance value for the iteration. Default: 1e-4}

\item{tas:}{vector of air temperature in degC}

\item{dewp:}{vector of dew point temperature in degC}

\item{wind:}{vector of wind speed in m/s}

\item{radiation:}{vector of solar shortwave downwelling radiation in W/m2}

\item{dates:}{vector with dates in format yyyy-mm-dd H:M:S, same length as the other inputs. If format yyyy-mm-dd is provided, the default hour is 12:00:00.}

\item{lon:}{value for the longitude of the location.}

\item{lat:}{value for the latitude of the location.}

\item{noNAs:}{logical, should \code{tas >= dewp} be enforced by swapping?}
}
\value{
A list of:

$data: wet bulb globe temperature in degC

$Tnwb: natural wet bulb temperature (Tnwb) in degC

$Tg: globe temperature in degC
}
\description{
Calculation of wet bulb globe temperature from air temperature, dew point temperature, radiation and wind.
}
\details{
This corresponds to the implementation for outdoors or in the sun conditions. Original fortran code by James C. Liljegren, translated by Bruno Lemke into Visual Basic (VBA) and Ana Casanueva into R.
}
\examples{
\dontrun{
# load the meteorological variables for example data in Salamanca:
data(eca_salam_jja_2003.Rdata)
wbgt.outdoors <- wbgt.Liljegren(eca_salam_jja_2003$tasmean, eca_salam_jja_2003$dewp, eca_salam_jja_2003$wind, eca_salam_jja_2003$solar, eca_salam_jja_2003$Dates, -5.66, 40.96)
}
}
\author{
A.Casanueva (21.02.2017).
}

