% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wbt.Stull.R
\name{wbt.Stull}
\alias{wbt.Stull}
\title{Calculation of wet bulb temperature, following Stull's method.}
\usage{
wbt.Stull(tas, hurs)
}
\arguments{
\item{tas:}{vector of air temperature in degC.}

\item{hurs:}{vector of relative humidity in \%.}
}
\value{
Wet bulb temperature in degC.
}
\description{
Calculation of wet bulb temperature from temperature and relative humidity.
}
\details{
Formulation from Stull 2011, Journal of Applied Meteorology and Climatology.
}
\examples{
\dontrun{
# load the meteorological variables for example data in Salamanca:
data("eca_salam_jja_2003.Rdata")
wbt <- wbt.Stull(eca_salam_jja_2003$tasmean, eca_salam_jja_2003$hurs)
}
}
\author{
A.Casanueva (15.08.2016).
}

