% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calZenith.R
\name{calZenith}
\alias{calZenith}
\title{Calculate zenith angle in degrees.}
\usage{
calZenith(dates, lon, lat, hour = FALSE)
}
\arguments{
\item{dates}{vector of dates in the format yyyy-mm-dd (or 'yyyy-mm-dd HH:MM:SS').  If format yyyy-mm-dd is provided, the default hour is 12:00:00.}

\item{lon}{value for the longitude of the location.}

\item{lat}{value for the latitude of the location.}

\item{hour}{logical. If TRUE, allow working with hourly data for zenith angle
calculation. Default: FALSE (12 UTC is used).}
}
\value{
zenith angle in degrees
}
\description{
Calculate zenith angle in degrees.
}
\examples{
\dontrun{ 
calZenith("1981-06-15",  -5.66, 40.96)
calZenith("1981-06-15 10:00:00",  -5.66, 40.96, hour=TRUE)
}

}
\author{
Anke Duguay-Tetzlaff, Translated to R by Ana Casanueva (17.01.2017)
}
