% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatStress.R
\docType{package}
\name{HeatStress}
\alias{HeatStress}
\alias{HeatStress-package}
\title{HeatStress}
\description{
Calculate heat stress indices
}
\details{
The package \code{HeatStress} provides methods to calculate heat stress indices.



The following calculation methods are implemented:
 \itemize{
 
 \item \code{wbt.Stull}: Calculation of the Wet Bulb Temperature (Stull 2011).
 \item \code{wbgt.Bernard}: Calculation of the Wet Bulb Globe Temperature in the shade (Bernard et al. 1999) 
 \item \code{wbgt.Liljegren}: Calculation of the Wet Bulb Globe Temperature in the sun (Liljegren et al. 2008) 
 \item \code{swbgt}: Calculation of the simplified wet bulb globe temperature (Buzan et al. 2015 and references therein).
 \item \code{apparentTemp}: Calculation of the apparent temperature (Buzan et al. 2015 and references therein).
 \item \code{effectiveTemp}: Calculation of the effective temperature (Coccolo et al. 2016 and references therein).
 \item \code{humidex}: Calculation of the humidex (Buzan et al. 2015 and references therein).
 \item \code{discomInd}: Calculation of the discomfort index (Coccolo et al. 2016 and references therein).
 \item \code{hi}: Calculation of the heat index (Buzan et al. 2015 and references therein).
 }
 
 Check the details of the indices and input variables: 

\code{library(HeatStress)}

\code{indexShow()}
}
