% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{Calculation of the heat index.}
\usage{
hi(tas, hurs)
}
\arguments{
\item{tas}{vector of air temperature in degC.}

\item{hurs}{vector of relative humidity in \%.}
}
\value{
Heat index in degrees Fahrenheit.
}
\description{
Calculation of the heat index from temperature and relative humidity.
}
\details{
Formula based on air temperature and relative humidity, following Rothfusz 1990 (National Weather Service Technical Attachment, SR 90-23), but adapted for degrees Celsius. This implementation includes some adjustments for high and low relative humidity values. Also, the original formula is not appropriate for low temperatures and heat index values. In those cases, a simpler formula is applied to calculate values consistent with Steadman's results. See: https://www.wpc.ncep.noaa.gov/html/heatindex_equation.shtml and https://github.com/ecmwf/thermofeel/blob/master/thermofeel/thermofeel.py#L782
}
\examples{
\dontrun{ 
# load the meteorological variables for example data in Salamanca:
data("data_obs") 
heatindex <- hi(data_obs$tasmean, hurs=data_obs$hurs)
}

}
\author{
A.Casanueva (22.03.2018). Modified in 12.08.2025.
}
