% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbgt.Liljegren.R
\name{wbgt.Liljegren}
\alias{wbgt.Liljegren}
\title{Calculation of wet bulb globe temperature, following Liljegren's method.}
\usage{
wbgt.Liljegren(tas, dewp, wind, radiation, dates, lon, lat, tolerance = 1e-04,
  noNAs = TRUE, swap = FALSE, hour = FALSE)
}
\arguments{
\item{tas}{vector of temperature in degC.}

\item{dewp}{vector of dewpoint temperature in degC.}

\item{wind}{vector of wind speed in m/s.}

\item{radiation}{vector of solar shortwave downwelling radiation in W/m2.}

\item{dates}{vector of dates in the format yyyy-mm-dd (or 'yyyy-mm-dd HH:MM:SS').  If format yyyy-mm-dd is provided, the default hour is 12:00:00.}

\item{lon}{value for the longitude of the location.}

\item{lat}{value for the latitude of the location.}

\item{tolerance}{(optional) tolerance value for the iteration. Default: 1e-4.}

\item{noNAs}{logical, should NAs be introduced when dewp>tas? If TRUE specify how to deal in those cases (swap argument)}

\item{swap}{logical, should \code{tas >= dewp} be enforced by swapping? Otherwise, dewp is set to tas. This argument is needed when noNAs=T.}

\item{hour}{logical. If TRUE, allow working with hourly data for zenith angle
calculation. Default: FALSE (12 UTC is used).}
}
\value{
A list of:

$data: wet bulb globe temperature in degC

$Tnwb: natural wet bulb temperature (Tnwb) in degC

$Tg: globe temperature in degC
}
\description{
Calculation of wet bulb globe temperature from air temperature, dew point temperature, radiation and wind.
}
\details{
This corresponds to the implementation for outdoors or in the sun conditions (Liljegren et al. 2008). Original fortran code by James C. Liljegren, translated by Bruno Lemke into Visual Basic (VBA) and Ana Casanueva into R.
}
\examples{
\dontrun{ 
# load the meteorological variables for example data in Salamanca:
data("data_obs") 
wbgt.outdoors <- wbgt.Liljegren(tas=data_obs$tasmean, dewp=data_obs$dewp, 
wind=data_obs$wind, radiation=data_obs$solar, dates= data_obs$Dates, lon=-5.66, lat=40.96)
}

}
\author{
A.Casanueva (21.02.2017).
}
